+++ Sources/ZcashLightClientKit/Synchronizer/CombineSDKSynchronizer.swift
@@ -279,6 +279,12 @@ extension CombineSDKSynchronizer: CombineSynchronizer {
         }
     }
 
+    public func transactionSubmissionStatus(for rawID: Data) -> SinglePublisher<TransactionSubmissionStatus?, Never> {
+        AsyncToCombineGateway.executeAction() {
+            await self.synchronizer.transactionSubmissionStatus(for: rawID)
+        }
+    }
+
     public func refreshExchangeRateUSD() {
         synchronizer.refreshExchangeRateUSD()
     }